feat(spec): a cold-start gate for n-gram drafting, and the split that made room
*** GPU GATES NOT RUN: the card was withheld by the user mid-change. This commit
*** is local and unpushed. make dev and make dev-test (CPU) are green and all
*** static gates pass; make test-gpu and verify-fast are OUTSTANDING, and the
*** A/B below has one confounded row that needs re-measuring.

Prompt-lookup drafting is cold by construction: it matches a suffix of the
generated text against earlier occurrences, so until the generation is long
enough there is nothing to match. Measured on Qwen3-14B Q6_K, the north-star
model, one request, same prompt:

  generated   accepted   tok/verify
    128         0.0 %       0.00     drafter never fires
    256         6.2 %       2.00     fires and LOSES
    512        39.6 %       7.33     strongly profitable
   1024        36.1 %       6.78

A verify costs ~50 ms against a ~6 ms decode step on this checkpoint, so 2.0
tokens per verify is four times under water while 6.78 pays. The existing
economics guard cannot catch this: it arms on spec_verifies >= 8 PER REQUEST and
a 128-token request produces about one, so the verdict is never made rather than
made badly - and a better threshold does not help something that never arms.

speculative.min_history conditions on tokens generated SO FAR, which is known at
the decision point, rather than on the request's eventual length, which is not.
Default 0 (off) pending the measurement below.

A/B at min_history=256, and the honest state of it:

  max_tokens=1024   default 153.06 tok/s (36.1 %)   gate 154.63 tok/s (31.2 %)
  max_tokens=512    both arms drafted 0 (the bimodality, LIMITATIONS.md)
  max_tokens=256    default 86.29 tok/s  <- NOT CREDIBLE, the same arm measured
                    153.70 minutes earlier; the card was contended. Re-measure.

So the warm case is not regressed (154.63 against 153.06, acceptance falling
36.1 -> 31.2 as the cold verifies are skipped, which is the mechanism working).
The cold case is unmeasured. The default stays 0 until it is not.

THE SPLIT. Adding the gate put engine_spec_ngram.cpp at 801 code LOC against a
hard-review ceiling of 800 - the file was exactly full. Allowlisting it to fit a
one-line feature would be the exemption creep the gate exists to prevent, so
instead the device-buffer lifecycle moved to engine_spec_buffers.cpp:
ensure_spec_buffers_, ensure_spec_state_scratch_, recurrent_slot_for_,
free_spec_buffers_.

Split on RESPONSIBILITY, not size - that file held the buffers the speculative
paths allocate and release, and the policy and execution that use them. The size
gate reported a conflation rather than a length. Pure move, no logic change:
801 -> 677 code LOC, and the I1 allowlist entry moved with its 13 allocation
sites (the gate failed on the REMOVAL first, which is the direction nobody
expects and the reason it is a two-way ratchet).
