+++ CMakeLists.txt
@@ -308,6 +308,7 @@ set(IMP_RUNTIME_SOURCES
     src/runtime/engine_scheduler.cpp
     src/runtime/engine_qwen3vl.cpp
     src/runtime/engine_spec_ngram.cpp
+    src/runtime/engine_spec_buffers.cpp
     src/runtime/engine_spec_capture.cpp
     src/runtime/engine_spec_mtp.cpp
     src/runtime/engine_sampling_stop.cpp
