+++ src/runtime/config.cpp
@@ -322,6 +322,7 @@ bool apply_one(RuntimeConfig& cfg, const std::string& dotted_key, const std::str
     B("speculative.verify_decode_attn", cfg.speculative.verify_decode_attn);
     B("speculative.verify_nvfp4_gemm", cfg.speculative.verify_nvfp4_gemm);
     B("speculative.verify_row_parity", cfg.speculative.verify_row_parity);
+    I("speculative.min_history", cfg.speculative.min_history);
     B("speculative.moe", cfg.speculative.moe);
     I("speculative.k", cfg.speculative.k);
     B("speculative.token_recycling", cfg.speculative.token_recycling);
